Spec'd out BeardBus substantially more.  Not a lot of concrete implementation yet, but we've got most of the major methods signatures figured out now I think.
Although I wonder if I might not be Doing It Wrong with pipes again.  It's always so tempting to make a pipe and then hand out an endpoint of it when first designing a new system, but it always seems to turn out to be much more powerful if you let the user create whatever they want and then hand you the opposite endpoint of what you would have given them.  In this case letting the caller of bind(*) give us a WriteHead would remove the need for the awkward inversion that is SimpleReactor, and would also let us do noteworthy things like aggregating the events from multiple bind calls into a single pipe.  On the other hand, it would screw up my current layout for unbinding, and I really do think that 99.998% of use cases will be one handler logic unit intended per bind call in which case pushing more code that's almost always boilerplate into the purview of the library user would be a bad.