BeardBus is now implemented... except routing completely doesn't work.
There's some lovely debugging calls showing that event translation itself is working.  But I found out that there's a BrowserSideObject class, and there's a MessagePassingJSObject class... and the latter is ephemeral (I get a new one from every js call even if it refers to the same js thing), and the latter is what I'm exposed to... so I can't use it as a pointer key for maps.  Bah.